Moved plain logic to EditboxPlain
