+++ src/frontend/components/edit/Editbox.svelte
@@ -2,8 +2,7 @@
     import { onMount } from "svelte"
     import { uid } from "uid"
     import type { Item, Line } from "../../../types/Show"
-    import { activeEdit, activePopup, activeShow, dictionary, os, outputs, overlays, redoHistory, refreshListBoxes, selected, showsCache, templates, variables } from "../../stores"
-    import { menuClick } from "../context/menuClick"
+    import { activeEdit, activePopup, activeShow, os, overlays, redoHistory, refreshListBoxes, selected, showsCache, templates, variables } from "../../stores"
     import Cam from "../drawer/live/Cam.svelte"
     import Image from "../drawer/media/Image.svelte"
     import Icon from "../helpers/Icon.svelte"
@@ -15,7 +14,7 @@
     import { addToPos } from "../helpers/mover"
     import { _show } from "../helpers/shows"
     import { getStyles } from "../helpers/style"
-    import Button from "../inputs/Button.svelte"
+
     import DynamicEvents from "../slide/views/DynamicEvents.svelte"
     import ListView from "../slide/views/ListView.svelte"
     import Mirror from "../slide/views/Mirror.svelte"
@@ -24,11 +23,12 @@
     import Visualizer from "../slide/views/Visualizer.svelte"
     import Website from "../slide/views/Website.svelte"
     import Clock from "../system/Clock.svelte"
-    import Movebox from "../system/Movebox.svelte"
+
     import { getAutoSize, getMaxBoxTextSize } from "./scripts/autoSize"
     import { addChords, chordMove } from "./scripts/chords"
     import { getLineText, getSelectionRange, setCaret } from "./scripts/textStyle"
     import { EditboxHelper } from "./EditboxHelper"
+    import EditboxPlain from "./EditboxPlain.svelte"
 
     export let item: Item
     export let filter: string = ""
@@ -248,7 +248,9 @@
 
     function getStyle() {
         if (!plain && $activeEdit.slide === null) return
-        html = EditboxHelper.getSyleHtml(item, plain)
+        let result = EditboxHelper.getSyleHtml(item, plain, currentStyle)
+        html = result.html
+        currentStyle = result.currentStyle
         previousHTML = html
     }
 
@@ -480,50 +482,6 @@
         return newLines
     }
 
-    // bindings
-    function removeBindings() {
-        menuClick("bind_item")
-    }
-
-    // actions
-    function removeAction(action) {
-        // TODO: this is a duplicate of SetTime and other places
-        let layoutRef: any[] = _show().layouts("active").ref()[0] || []
-        let slideRef: any = layoutRef[$activeEdit.slide!] || {}
-        let slideItems = _show().get("slides")?.[slideRef.id]?.items || []
-
-        if ($activeEdit.id) getItems()
-        function getItems() {
-            let slide = {}
-            if ($activeEdit.type === "overlay") slide = $overlays
-            else if ($activeEdit.type === "template") slide = $templates
-
-            slideItems = slide[$activeEdit.id!]?.items
-        }
-
-        if (!slideItems) return
-
-        let actions = clone(slideItems[index].actions)
-        delete actions[action]
-
-        if ($activeEdit.type === "overlay" || $activeEdit.type === "template") {
-            history({
-                id: "UPDATE",
-                oldData: { id: $activeEdit.id },
-                newData: { key: "items", subkey: "actions", data: [actions], indexes: [index] },
-                location: { page: "edit", id: $activeEdit.type + "_items", override: "deleteitemaction_" + index },
-            })
-
-            return
-        }
-
-        history({
-            id: "setItems",
-            newData: { style: { key: "actions", values: [actions] } },
-            location: { page: "edit", show: $activeShow!, slide: slideRef.id, items: [index], override: "deleteitemaction_" + slideRef.id + "_items_" + index },
-        })
-    }
-
     // timer
     let today = new Date()
     setInterval(() => (today = new Date()), 1000)
@@ -585,12 +543,6 @@
     // $: autoSize = height < width ? height / 1.5 : width / 4
     // $: autoSize = Math.min(height, width) / 2
     // $: autoSize = getAutoSize(item)
-
-    const actions = {
-        transition: { label: "popup.transition", icon: "transition" },
-        showTimer: { label: "actions.show_timer", icon: "time_in" },
-        hideTimer: { label: "actions.hide_timer", icon: "time_out" },
-    }
 
     // CHORDS
 
@@ -692,42 +644,7 @@ bind:offsetWidth={width} -->
     on:mousedown={mousedown}
 >
     {#if !plain}
-        <Movebox {ratio} active={$activeEdit.items.includes(index)} />
-
-        <div class="actions">
-            <!-- bindings -->
-            {#if item.bindings?.length}
-                <div title={$dictionary.actions?.remove_binding} class="actionButton" style="zoom: {1 / ratio};left: 0;right: unset;">
-                    <Button on:click={removeBindings} redHover>
-                        <Icon id="bind" white />
-                    </Button>
-                    {#if item.bindings.length > 1}
-                        <span class="actionValue">{item.bindings.length}</span>
-                    {:else}
-                        <span class="actionValue">
-                            {#if item.bindings[0] === "stage"}
-                                <T id="menu.stage" />
-                            {:else}
-                                {$outputs[item.bindings[0]]?.name}
-                            {/if}
-                        </span>
-                    {/if}
-                </div>
-            {/if}
-            <!-- actions -->
-            {#if item.actions}
-                {#each Object.keys(item.actions) as action}
-                    <div title={actions[action] ? $dictionary[actions[action].label.split(".")[0]]?.[actions[action].label.split(".")[1]] : ""} class="actionButton" style="zoom: {1 / ratio};left: 0;right: unset;">
-                        <Button on:click={() => removeAction(action)} redHover>
-                            <Icon id={actions[action]?.icon} white />
-                        </Button>
-                        {#if typeof item.actions[action] === "number"}
-                            <span class="actionValue">{item.actions[action]}s</span>
-                        {/if}
-                    </div>
-                {/each}
-            {/if}
-        </div>
+        <EditboxPlain {item} {index} {ratio} />
     {/if}
     {#if item?.lines}
         <!-- TODO: remove align..... -->
@@ -847,31 +764,6 @@ bind:offsetWidth={width} -->
         font-size: var(--auto-size) !important;
     }
 
-    .actions {
-        position: absolute;
-        top: 0;
-        left: 0;
-
-        display: flex;
-        flex-direction: column;
-    }
-    .actionValue {
-        font-size: small;
-        opacity: 0.5;
-        font-weight: bold;
-        padding: 0 5px;
-        text-shadow: none;
-    }
-    .actionButton {
-        display: flex;
-        align-items: center;
-        background-color: var(--focus);
-        color: var(--text);
-    }
-    .actionButton :global(button) {
-        padding: 5px !important;
-        z-index: 3;
-    }
     .chords :global(.chord) {
         position: absolute;
         transform: translateY(-100%);
